refactor: move util functions to a separate file, remove old MoneyRequestAmountPage
